+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## dev
 
+- Use `Iterable[metadata.Distribution]` instead of `List[metadata.Distribution]` in `venv_inspect.py` (#629)
 - Make usage message in `pipx run` show `package_or_url`, so extra will be printed out as well
 - Add `--force-reinstall` to pip arguments when `--force` was passed
 - Use the py launcher, if available, to select Python version with the `--python` option